game.js -> index.ts
